requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.2
Pillow==5.2.0
progressbar==2.5
pyparsing==2.2.1
python-dateutil==2.7.3
scipy==1.1.0
six==1.11.0
torch==0.4.1
torchvision==0.2.1
